leaudio: Fix build warnings in logs
packages/modules/Bluetooth/system/bta/le_audio/client.cc:3840:13: warning: format specifies type 'unsigned short' but the argument has type 'LeAudioContextType' [-Wformat]
 3838 |     dprintf(fd, "  configuration: %s  (0x%08hx)\n",
      |                                          ~~~~~
 3839 |             bluetooth::common::ToString(configuration_context_type_).c_str(),
 3840 |             configuration_context_type_);
      |             ^~~~~~~~~~~~~~~~~~~~~~~~~~~
      |             static_cast<unsigned short>( )
packages/modules/Bluetooth/system/bta/le_audio/client.cc:3850:15: warning: format specifies type 'int' but the argument has type 'value_type' (aka 'bluetooth::le_audio::UnicastMonitorModeStatus') [-Wformat]
 3849 |       dprintf(fd, "  Local sink notified state: %d\n",
      |                                                 ~~
 3850 |               sink_monitor_notified_status_.value());
      |               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
      |               static_cast<int>(                    )

Bug: 331775328
Test: mmm packages/modules/Bluetooth
Flag: Exempt, trivial fix, log only

+++ system/bta/le_audio/client.cc
@@ -3837,7 +3837,7 @@ class LeAudioClientImpl : public LeAudioClient {
                  : "Targeted Announcements"));
     dprintf(fd, "  configuration: %s  (0x%08hx)\n",
             bluetooth::common::ToString(configuration_context_type_).c_str(),
-            configuration_context_type_);
+            static_cast<uint16_t>(configuration_context_type_));
     dprintf(fd, "  local source metadata context type mask: %s\n",
             local_metadata_context_types_.source.to_string().c_str());
     dprintf(fd, "  local sink metadata context type mask: %s\n",
@@ -3847,7 +3847,7 @@ class LeAudioClientImpl : public LeAudioClient {
             sink_monitor_mode_ ? "true" : "false");
     if (sink_monitor_notified_status_) {
       dprintf(fd, "  Local sink notified state: %d\n",
-              sink_monitor_notified_status_.value());
+              static_cast<int>(sink_monitor_notified_status_.value()));
     }
     dprintf(fd, "  Source monitor mode: %s\n",
             source_monitor_mode_ ? "true" : "false");
